Extract useEditorURL helper hook to read the editor url
